Allow ENOSPC and EIO for reads and writes.
EIO (unfortunately) commonly happens while stress testing lustre. That is
not nice, but for now we have to live with that. It probably better should
become and option to ignore this error. ENOSPC will easily come up when the
filesystem fills up and when several indepedent instances of the program
(and also other test programs) compete for the remaining disk space.
So we almost silently ignore this error.